requirements.txt
requests
pycryptodome
pypiwin32
colorama
clipboard
